Minor refactor for package reorg
