+++ tests/doc_db_inserter/doc_db_inserter_test.py
@@ -1,7 +1,7 @@
 import unittest
 from uuid import uuid4
 
-from flowcept.commons.doc_db.document_db_dao import DocumentDBDao
+from flowcept.commons.daos.document_db_dao import DocumentDBDao
 
 
 class TestDocDBInserter(unittest.TestCase):
