+++ tests/plugins/test_dask.py
@@ -6,8 +6,8 @@ import numpy as np
 
 from dask.distributed import Client
 
-from flowcept.commons.doc_db.document_db_dao import DocumentDBDao
-from flowcept.commons.doc_db.document_inserter import (
+from flowcept.commons.daos.document_db_dao import DocumentDBDao
+from flowcept.flowceptor.consumers.document_inserter import (
     DocumentInserter,
 )
 from flowcept.commons.flowcept_logger import FlowceptLogger
